reviews app added, data transferred to new app, minor permissions changes
