+++ api_yamdb/api/permissions.py
@@ -1,5 +1,6 @@
 from rest_framework.permissions import BasePermission, SAFE_METHODS
 
+
 class AdminOrRead(BasePermission):
 
     def has_permission(self, request, view):
