+++ api_yamdb/users/admin.py
@@ -1,6 +1,9 @@
 from django.contrib import admin
 from .models import User
 
+
 class AuthorAdmin(admin.ModelAdmin):
     pass
+
+
 admin.site.register(User, AuthorAdmin)
